pyproject.toml
[project]
name = "ct-images-augmentation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "mateuszGorczany", email = "mateusz.g64@interia.pl" }
]
dependencies = [
    "wandb>=0.17.0",
    "torch>=2.3.0",
    "monai>=1.3.0",
    "transformers>=4.40.2",
    "lightning>=2.2.4",
    "pydantic>=2.7.1",
    "hydra-core>=1.3.2",
    "dacite>=1.8.1",
    "matplotlib>=3.9.0",
    "nibabel>=5.2.1",
    "scikit-image>=0.24.0",
    "torchvision>=0.18.1",
    "monai-generative>=0.2.3",
    "lpips>=0.1.4",
    "fastmri>=0.3.0",
    "einops>=0.8.0",
    "einops-exts>=0.0.4",
    "rotary-embedding-torch>=0.6.4",
    "timm>=1.0.8",
    "pytorch-msssim>=1.0.0",
    "torchmetrics[image]>=1.4.2",
]
readme = "README.md"
requires-python = "== 3.10.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.4",
    "streamlit>=1.34.0",
    "tensorboard>=2.17.0",
    "pylyzer>=0.0.61",
    "pyrasite>=2.0",
    "pyright>=1.1.379",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
